Migration: Change Expert<->>Users to Experts<<->>Users
This is actually a workaround: most users (conseillers) are expected to belong to one expert (équipe référent). However, some (référents) have different set of (compétences) for different (territoires), and as of today, exist in duplicate entries in the expert (référents) table. As we want to link a single user (conseiller) to all these experts, we need a many-to-many association.