Manually add the template.data.gouv.fr main css
(this is version 1.1.8 apparently)

I’m not really satisfied about it. The css file itself can be @imported in a regular rails sass file, or referenced from a stylesheet_tag in the header; however, it references the font-faces with their regular paths ("../fonts/Evolventa/...") which are now how they would be served by the assets pipeline. I ended up putting the font files in the /public folder.

I hope to clean this all when modernizing the assets loading, and converting the whole site to the betagouv template.