Validate diagnosis attributes at step 4
Some diagnoses have no happened_on (which is why I’m sorting by created_at in _diagnoses.haml).

This happens when the user passes invalid data in step 3 and the browser doesn’t block them. This conditionally validates the presence of :happened_on, as well as :visitee for good measure.