Fix date formatting for IE11 for diagnosis date
IE11 adds `left-to-right` marks around the `/`, which then breaks the date parsing on ruby’s side. Let’s just format it manually.

This is a quick and dirty fix, until we
- use a proper date library
- refactor this form to not use the same formatter for display and for data. 😓

This is ugly, but not as bad as it seems. As a matter of fact, we require french formatting when used as text input, and the placeholder reads 'jj/mm/aaaa'.