Add a honeypot field to the solicitation form
We don’t have a spam problem — yet. This is a basically free solution; if/when it’s not enough, we’ll start looking for more.

Note: this gem adds inline CSS, that _might_ be an issue for CSP. We’ll see. In the worse case, it’s not really hard to add a honeypot field ourselves.