Use :validatable for Users
Instead of validating emails and passwords ourselves.
Override password_required? to allow empty password before the invitation is accepted.

This is already the behaviour we were aiming for. However, using `super` in `password_required?` fixes an issue where we couldn’t modify users in /admin after they accepted the invitation. This is because `database_authenticatable` overrides the :password attr_reader for obvious reasons, which means our implementation of `password_required?` was wrong. The implementation in `:database_authenticatable` correctly requires a password only when changing the password :)