Use a png logo for BDF
This is unfortunate. In #832 we changed the BDF logo, but the old version was png and the new one jpeg. However, in institution_image() we resolve paths using png in priority. Scalingo still has the old version cached, which ends up served instead of the new (jpeg) one.

Let’s just use png and fight this another day.

(See also #644 😩)