Display foreign key violations errors in /admin when trying to destroy
When trying to destroy an object that is still referenced, for example an Institution that still has Antennes, PG will throw an exception and AA will fail and return an `error 500` blank page.

This catches and displays the underlying error; there seems to be no mechanism in AA to handle destroy failures at all. 🤷🏻‍♂️