Merge the subjects of individual users in a team
When importing users of a same team, the subjects are set on the team. Typically, partners will provide csv with each member of the team having only their subject selected. The team must have all the subjects of its members.

This also means that when importing users, we only add subjects, never remove them.

fixes #1310