Remove some dynamic magic for User and Expert options in activeadmin
Change the way the flags are set and cleared in admin for Users and Experts.

Instead of dynamically building batch and item actions, just build one batch action with form_options. It’s not used very often, and it’s less fragile.

(Done in passing for #1348; this isn’t really related but I was there.)