Make sure Need#update_status works correctly after the matches are made
Fixes #1421

The `update(status:...)` actually failed, because of a validation error in matches. It was due to this validation in Match:

```
validates :expert, uniqueness: { scope: :need_id, allow_nil: true }
```
This code is called in a weird context: we’re actually updating diagnosis params from StepController#update_matches, and the matches `touch`es back the Need. It seems that Need.update tries to validate the in-memory matches objects, but fails because the same data is already in database, and it doesn’t know it’s actually the same record.

Anyway. Nothing that a .reload can’t fix. 🤷‍♀️