Remove invalid old relations in User
User.has_many :antenne_territories and User.has_many :antenne_regions do not exist anymore, the equivalent in Expert are already commented out.

Clicking the link to matches from /admin/users does not use these relations directly, but the introspection in User to find the inverse crashed because of these.